Abstracting desktop sources
Most are the same, only varying parts should be varying.
