+++ zeapp/badge/src/linuxMain/kotlin/de/berlindroid/zekompanion/BadgeManager.kt
@@ -1,30 +1,13 @@
 package de.berlindroid.zekompanion
 
-import java.io.File
-import java.lang.RuntimeException
+import com.fazecast.jSerialComm.SerialPort
 
-class AppleBadgeManager : BadgeManager {
-    override suspend fun sendPayload(payload: BadgePayload): Result<Int> = if (isConnected()) {
-        val serial = getUsbModems().last()
-        val bytes = payload.toBadgeCommand().toByteArray()
-        serial.writeBytes(bytes)
-        Result.success(bytes.size)
-    } else {
-        Result.failure(RuntimeException("No badge connected."))
-    }
-
-    override fun isConnected(): Boolean {
-        return getUsbModems().size == 2
-    }
-
-    private fun getUsbModems() = File("/dev/")
-        .listFiles()
-        .orEmpty()
-        .filter {
-            it.name.startsWith("ttyACM")
-        }
-}
 
 actual typealias Environment = Any
 
-actual fun buildBadgeManager(environment: Environment): BadgeManager = AppleBadgeManager()
+actual fun buildBadgeManager(environment: Environment): BadgeManager = object : JvmBadgeManager() {
+    override fun getBadger2040s() = SerialPort.getCommPorts().filter {
+        it.descriptivePortName.contains("Badger") &&
+                it.systemPortPath.contains("tty")
+    }.toList()
+}
